Replace dropdown target selection with list
